[store]: refactor: adopt action-handler proposed by @hulunbier
