+++ fusestore/store/src/executor/mod.rs
@@ -1,0 +1,7 @@
+// Copyright 2020-2021 The Datafuse Authors.
+//
+// SPDX-Lise-Identifier: Apache-2.0.
+
+mod action_handler;
+
+pub use action_handler::ActionHandler;
